(Refact) seperate profileupdate
